requirements.txt
gitdb==4.0.11
GitPython==3.1.42
psutil==5.9.8
python-dotenv==1.0.1
schedule==1.2.1
Send2Trash==1.8.2
smmap==5.0.1
